retract symbols that are not in stable
since unstable was merged into vivid_archive when releasing a stable update
for plasma there's some pollution going on as to what belongs where.
these symbols originally were introduced in the unstable branch and
eventually are failing the stable branch now due to the merge and not
being available. so, retract them again